misc/file-magic/solve/requirements.txt
pycryptodome==3.15.0
Pillow==9.2.0
pwntools==4.8.0
